# backend-go/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.loginRequest:
    properties:
      captcha_code:
        description: 用户输入的验证码答案
        type: string
      captcha_token:
        description: 验证码 JWT（与 captcha_code 配对）
        type: string
      password:
        type: string
      username:
        type: string
    type: object
...
  handlers.modemUpdate:
    properties:
      alias:
        type: string
    type: object
...
  models.Modem:
    properties:
      access_technologies:
        description: 接入技术，如 lte
        type: string
      alias:
        description: 用户自定义别名
        type: string
      connection_duration:
        description: 秒
        type: integer
      created_at:
        type: string
      current_bands:
        type: string
      current_modes:
        type: string
      device_path:
        description: 设备节点，如 ttyUSB0
        type: string
      firmware_revision:
        type: string
      hardware_revision:
        type: string
      iccid:
        type: string
      id:
        type: integer
      imei:
        description: IMEI，可能为空
        type: string
      imsi:
        type: string
      is_active:
        description: false 表示设备已拔出
        type: boolean
      last_seen:
        type: string
      manufacturer:
        type: string
      mm_object_path:
        description: 唯一键
        type: string
      model:
        type: string
      operator:
        description: 运营商名称
        type: string
      phone_number:
        type: string
      plugin:
        description: mmcli 插件名
        type: string
      ports:
        type: string
      registration_state:
        type: string
      rx_bytes:
        type: integer
      signal_quality:
        description: 信号质量 0-100
        type: integer
      sim_operator_code:
        description: MCC+MNC
        type: string
      sim_operator_name:
        description: SIM 卡归属运营商名
        type: string
      status:
        type: string
      tx_bytes:
        type: integer
    type: object
  models.Notification:
    properties:
      audience:
        description: 受众：admin/support/all/user
        type: string
      body:
        description: 通知正文
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_read:
        type: boolean
      target_user_id:
        description: audience=user 时指定目标用户
        type: integer
      title:
        type: string
      type:
        description: 通知类型，如 sms_failed
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: 多USB 4G Modem 管理系统 API
  title: SimNexus API
  version: "1.0"
paths:
  /api/v1/admin/logs/stream:
    get:
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE流
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 实时日志SSE流（仅管理员）
      tags:
      - 系统
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: 验证用户名/密码（可选验证码），返回 JWT
      parameters:
      - description: 登录信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: 用户登录
      tags:
      - 认证
  /api/v1/auth/me:
    get:
      description: 返回当前登录用户及其 RBAC 角色
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取当前用户信息
      tags:
      - 认证
...
  /api/v1/modems/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Modem'
            type: array
      security:
      - BearerAuth: []
      summary: 获取用户有权限的设备列表
      tags:
      - 设备管理
  /api/v1/modems/{id}:
    get:
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Modem'
      security:
      - BearerAuth: []
      summary: 获取单个设备信息
      tags:
      - 设备管理
    patch:
      consumes:
      - application/json
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: integer
      - description: 修改字段
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.modemUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Modem'
      security:
      - BearerAuth: []
      summary: 修改设备属性
      tags:
      - 设备管理
  /api/v1/modems/{id}/detail:
    get:
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取设备详情（含短信统计）
      tags:
      - 设备管理
  /api/v1/modems/{id}/refresh:
    post:
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Modem'
      security:
      - BearerAuth: []
      summary: 手动刷新设备状态
      tags:
      - 设备管理
  /api/v1/modems/available:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Modem'
            type: array
      security:
      - BearerAuth: []
      summary: 获取所有设备（资源库）
      tags:
      - 设备管理
  /api/v1/notifications:
    get:
      parameters:
      - description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Notification'
            type: array
      security:
      - BearerAuth: []
      summary: 获取通知列表
      tags:
      - 通知
  /api/v1/notifications/{id}/read:
    post:
      parameters:
      - description: 通知ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 标记单条通知为已读
      tags:
      - 通知
  /api/v1/notifications/read-all:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 一键标记所有通知为已读
      tags:
      - 通知
  /api/v1/notifications/unread-count:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取未读通知数量
      tags:
      - 通知
...
securityDefinitions:
  BearerAuth:
    description: '格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"